Class for building the source
Very rough way of doing the total field scattered field method. It's simple when the incidence angle is 0 or 90, but it becomes more complex for other incident angles. I could find any good sources for how to do this as all of them are either too complex to actually let you understand the process or they tell you to read the same chapter from the same book that I can't get my hands on. I believe the only function from an outside class I used was the wave() function that is stored in the general_functions_class file. Some of the code I rushed Wednesday night so those parts may not be commented very well so don't be afraid to ask me any questions about it.